--- a/.github/workflows/ci-optimized.yml
+++ b/.github/workflows/ci-optimized.yml
@@ -1,10 +1,8 @@
-name: Optimized CI/CD
+name: Optimized CI/CD (DEPRECATED - Use Simple CI)
+# This workflow is deprecated. Please use .github/workflows/ci-simple.yml
 
 on:
-  push:
-    branches: [ main, develop, 'feature/**' ]
-  pull_request:
-    branches: [ main, develop ]
+  workflow_dispatch: # Only manual trigger
 
 # Cancel previous runs on the same branch
 concurrency: