--- a/src/components/PerformanceDashboard/index.ts
+++ b/src/components/PerformanceDashboard/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Performance Dashboard Components
+ * Export all dashboard components for easy import
+ */
+
+export { PerformanceDashboard } from './PerformanceDashboard';
+export { FPSChart } from './FPSChart';
+export { MemoryChart } from './MemoryChart';
+export { AudioLatencyChart } from './AudioLatencyChart';
+export { AlertsPanel } from './AlertsPanel';
+export { PerformanceStats } from './PerformanceStats';
+export { QualityControls } from './QualityControls';