… .devcontainer/README.md
… .devcontainer/devcontainer.json
… .devcontainer/docker-compose.yml
… .devcontainer/init-firewall.sh
… .github/workflows/ci-simple.yml
… .github/workflows/ci-staged.yml
… .github/workflows/ci.yml
… .github/workflows/deploy.yml
… .github/workflows/nextjs.yml
… CICD_IMPLEMENTATION_REPORT_2025.md
… CICD_IMPROVEMENT_PLAN_2025.md
… CLAUDE.md
… GITHUB_ACTIONS_ULTRATHINK_ANALYSIS.md
… PHASE1_TEST_STABILIZATION_REPORT.md
… PHASE1_ULTRATHINK_ANALYSIS.md
… PHASE2_DESIGN_REVIEW.md
… PHASE2_IMPLEMENTATION_STATUS.md
… PHASE2_ULTRATHINK_ANALYSIS.md
… PR_ULTRATHINK_REVIEW_2025.md
… README.md
… TEST_FAILURE_ANALYSIS.md
… V1Z3R_STRATEGIC_ROADMAP_2025.md
… docs/DEVCONTAINER_SETUP_2025.md
… docs/api/performance-monitor.md
… docs/features/PERFORMANCE_MONITORING_GUIDE.md
… src/__mocks__/webAudioMock.ts
… src/__mocks__/webGPUMock.ts
… src/components/PerformanceDashboard/AlertsPanel.tsx
… src/components/PerformanceDashboard/AudioLatencyChart.tsx
… src/components/PerformanceDashboard/FPSChart.tsx